// src/app/models/mrp/raw-material.ts
import { Material } from './material';

export class RawMaterial {
    id: number;
    roomId: number;
    room: string;
    name: string;
    productId: number;
    packageId: number;
    performance: number;
    barcode: string;
    retailPrice: number;
    actualCost: number;
    observation: string;
}